winewayland.drv: Introduce wayland driver
Implement a wayland driver with GDI support. This driver allows users to
run Windows applications directly on wayland compositors without an
intermediate layer to translate from X11 to wayland.

The driver is currently not built by default and needs to be enabled
with the '--with-wayland' configuration flag.

A major incompatibility between win32 and wayland is the absence (by
design) of any support for absolute screen window positioning in the
wayland protocol. For win32 transient windows (menus, tooltips etc) the
driver tries to work around the lack of absolute positioning by
anchoring them to an owning wayland surface and treating them as
subsurfaces of that owner. Screen coordinates for such windows are
transformed to local coordinates relative to the owning surface,
allowing correct placement through relative subsurface movement, which
is supported by wayland. Absolute positioning is not supported at this
time for non-transient top-level windows.

The wayland driver currently creates a single backing window surface for
each toplevel window, and all children windows use that single window
surface.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>