winewayland.drv: Support display mode changes
The wayland protocol doesn't allow clients to change the display mode.
We can emulate such changes by allowing clients (in this case Wine) to
draw at the desired resolution and then instructing the wayland
compositor to scale the result accordingly. The used scaling method
preserves the aspect ratio of the requested display mode.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>