winewayland.drv: Avoid indefinite blocks on GL swap buffers
The contents of GL or Vulkan windows are rendered on subsurfaces
with the parent surface used for the decorations. Such GL/VK
subsurfaces may want to commit their contents before the parent
surface has had a chance to commit. In such cases the GL/VK commit
will not be displayed, but, more importantly, will not get a frame
callback until the parent surface is also committed. Depending on
the presentation mode, a second GL/VK buffer swap may indefinitely
block waiting on the frame callback.

This commit introduces a function to ensure that a wayland surface
is mapped, by committing a dummy buffer if necessary. This function
is called before GL swap buffers operations to ensure the swap cannot
block indefinitely as described above.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>