winewayland.drv: Ensure fullscreen state is properly applied
In order for the fullscreen state, or a surface size change while
fullscreen to be applied by the wayland compositor, the clients are
required to commit a new buffer for the fullscreen toplevel.

There are some scenarios when the toplevel window surface contents are
not invalidated when transitioning to fullscreen, so the proper wayland
state is not set. Fix this by detecting such state transitions and
invalidating the window surface to ensure we eventually commit a new
buffer.

Such a scenario can occur with the upcoming Vulkan support, where the
the Vulkan window contents are implemented with a subsurface of the
toplevel decoration surface. During a fullscreen state change the
toplevel surface may not be updated, so we need to force an invalidation
as described above.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>