winewayland.drv: Use the native/HiDPI resolution for window rendering
Always use the native/HiDPI resolution for window rendering, even when
the surface backing the window is presented on a scaled wayland output.
This provides the full resolution to applications.

To achieve this we inform the wayland compositor that our surfaces are
already scaled for the output they are in. We update this scaling
information only when a surface moves *fully* to a new output. Until
this move happens we maintain the current scaling, e.g., when a surface
is between outputs.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>