winewayland.drv: Rework handling of transient toplevel windows
We use heuristics to decide which toplevels are to be considered
transient (and thus be backed by wayland subsurfaces): the toplevel
needs to be undecorated and have a reasonably small size relative
to its main monitor.

In some cases transient windows using global coordinates are created
once and then moved around and shown as needed. This means that we
cannot decide the effective parent only at creation time, but rather
keep track of the current effective parent to be able to place such
transient windows at the correct surface-local positions.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>