winewayland.drv: Send wine mode change message directly to threads
We currently notify each thread about wine mode changes by sending an
internal message to one of the thread's windows.

Going forward we won't be able to guarantee that all threads that need
to be notified about wine mode changes will have any windows. So,
instead of targeting a specific window, use the recently exposed
internal function to send the internal message to the thread directly.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>