winewayland.drv: Handle missing wayland toplevel for driver window messages
Some posted driver window messages assume the existence of a backing
wayland xdg_toplevel, which exists at the time of posting but may not
exist when the message is dispatched.

This can happen when between post and dispatch time the window changes
from a toplevel to become a child (through SetParent) or changes
ownership and is recreated as a wl_subsurface. This commit ensures we
handle this case gracefully, effectively ignoring the invalid messages.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>