winewayland.drv: Ensure we handle all necessary configure events
Checking only for a valid serial is not enough to guarantee that there
is a pending WM_WAYLAND_CONFIGURE message: we may have processed the
message but not acked the configure request due to surface size
incompatibilities (see window.c:update_wayland_state()). To solve this,
use a separate 'processed' flag to track WM_WAYLAND_CONFIGURE message
processing.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>