winewayland.drv: Avoid changing wayland state from non-window thread
In some cases, notably when an application calls UpdateLayeredWindow,
position and size changes may be emitted from a thread other than the
window thread. Since in the current implementation updating the wayland
state needs to happen in the context of the window thread to avoid racy
interactions, post a message to update the state in the right thread.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>