winewayland.drv: Add function to dispatch Wayland events.
Introduce a mechanism to dispatch Wayland events from a queue,
potentially also reading and queueing new events. This is similar to the
wayland-client wl_display_dispatch_queue() function, with the addition of
a timeout to wait for while waiting for new events to arrive.

This function also ensures to dispatch any per-process queue events with
the per-process instance lock held.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>