winewayland.drv: Read and dispatch Wayland events.
Create a separate reader thread to read Wayland events from the compositor
and dispatch them to their appropriate queues. The reader thread
notifies all Wayland-enabled threads of possible new events by using
an unnamed pipe for each thread. The read end of the pipe is
used as the Wine driver queue fd for the respective thread.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>