winewayland.drv: Refine heuristics for relative positioning of windows.
The current, simple heurestic for using a Wayland subsurface and relative
positioning for displaying a window is: it doesn't have a caption
AND it has an owner.

Since many applications use unowned top-level windows for various kinds
of popups, this commit adds a heurestic to detect detect such windows
and use relative positioning for them too: if a window is an unowned
top-level AND it doesn't have a caption AND it's relatively small
compared to the monitor size THEN use a subsurface/relative positioning.

Since for such windows we don't have a owner window, and thus a Wayland
surface parent to attach them to, we use the most recent surface the
user interacted with as the effective parent.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>