winewayland.drv: Update Wayland surface state.
When the properties of window change, update the state of the Wayland
surface backing that window. This potentially involves requesting a
state change from the compositor and setting proper values for the
surface geometry and subsurface position, when applicable.

In order to avoid protocol errors, apply the new state only if the
current or pending Wayland configure events are compatible with it. In
the latter case also acknowledge the pending event.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>