winewayland.drv: Handle Wayland surface output positioning.
Track the output(s) a Wayland surface is visible in, to allow us to
update the position of the associated window in the Windows virtual
screen space.

We always place windows at the top-left origin of their main output.
The main output is updated to be the leftmost output a surface has
entered. Support for more intuitive handling of windows with RTL layouts
remains to be done.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>