winewayland.drv: Use output scale on Wayland surface buffers.
Use the scale of the main output of each Wayland surface to always
render at the native resolution on that output. This effectively turns
off compositor-side scaling, giving applications/Wine complete control
over how to scale.

Note that the for non-cursor surfaces, we don't use
wl_surface_set_buffer_scale, but instead rely on wp_viewport to apply
the proper scaling. We do this because wl_surface_set_buffer_scale
requires buffer sizes to be integer multiples of the scale, but since
the buffer size corresponds to the window size, this is not a
requirement we can always uphold.

Although the same requirement applies to cursor surfaces, there are
fundamental issues with libwayland-cursor (see [1]) that block
compositors from strictly enforcing the spec. For now we can keep using
the buffer scale for cursors.

[1] https://gitlab.freedesktop.org/wayland/wayland/-/issues/194

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>