winewayland.drv: Assign unique 32-bit output ids.
Assign a unique 32-bit id to each output, based on its Wayland name,
which is guaranteed to be consistent across output reconnects. We do
this so that we can get an id that's consistent across all threads, and
which is small enough to use as a window message parameter in upcoming
commits.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>