winewayland.drv: Implement wglShareLists.
Sharing of display lists works differently in EGL and WGL. In case of
EGL it is done at context creation time but in case of EGL it can also
be done using wglShareLists.

We handle this by creating an EGL context in wglCreateContext /
wglCreateContextAttribsARB and when a program requests sharing we
recreate the destination context if it hasn't been made current and it
hasn't shared display lists before.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>