winewayland.drv: Implement wglSwapBuffers.
Before swapping we check that the target Wayland surface allows drawing,
to avoid commits with incorrectly sized buffers during reconfigurations.

We also ensure that the parent (decoration) Wayland surface is mapped to
avoid blocking: if the parent surface is not mapped the GL commit will
not be displayed, but, more importantly, will not get a frame callback.
Depending on the presentation mode, a second GL buffer swap could
indefinitely block waiting on the frame callback.

This commit implements unthrottled swapping behavior. The next commit
will add support for proper SwapInterval handling.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>