winewayland.drv: Support compositor-side scaling.
Introduce a mechanism to select whether we should use application-side
or composidor-side scaling. This is controlled by the "HiDPIScaling"
global or per-app driver option, either "Application" (the default) or
"Compositor".

With application-side scaling, which is the default, the application has
access to the full/native resolution of the outputs, and needs to manage
scaling manually (e.g., by using the Wine DPI options).

With compositor-side scaling, the application has access only to the
scaled resolution of the outputs, and the compositor automatically
scales the provided buffers.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>