winewayland.drv: Track the Wine output containing a window.
Track the output Wine considers to contain the window backed by a
Wayland surface. Transiently, this may be different from the output
Wayland considers to be the "main" one for this surface. This is useful
in case the Wayland compositor has not yet sent the relevant output
surface enter event(s).

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>