winewayland.drv: Maintain a per-process wayland instance.
We create a separate per-process instance, to be able to service win32
requests that do not require the presence of a GUI enabled thread and/or
require a single source of truth for all threads in a process (e.g.,
GetCurrentDisplaySettings).

All dispatches of Wayland events for per-process Wayland objects
should be done with the per-process instance lock held.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>