winewayland.drv: Experimental support for systray.
Use a standalone systray, adapted from winex11, to present icons for
applications.

Because the context menu on icons is drawn using global positioning
across process boundaries, we cannot currently anchor the menu window
properly to its effective parent (i.e., the icon), and we draw it as a
toplevel instead. Practically this has the effect that the menu window
will be placed at arbitrary locations on the screen by the Wayland
compositor.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>