winewayland.drv: Introduce abstraction for native buffers.
Introduce a structure to hold native buffers, which may consist
of a number of planes. The memory of each plane is represented
as a file descriptor, the exact type of which will depend on
the underlying buffer type. Future commits will add functions
to create native buffers for various buffer types.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>