winewayland.drv: Introduce support for DRI_PRIME.
Until now, users could only specify the DRM device that they want us to
use with the "DRMDevice" driver registry option, setting it to the
render node exposed by the GPU that they want to use (e.g.
"/dev/dri/renderD128").

This adds support to open a DRM device based on the value of the
DRI_PRIME environment variable. Users can set it to:

a. "1", and in such case we try to open the non-default device.
   The non-default device is a device that is different from
   whatever the compositor is using. But if the compositor didn't
   advertise dma-buf feedback, we won't be able to get this
   information. In such case, we are going to consider that the
   primary system GPU is the default device.

b. When users want to be more specific, they can set DRI_PRIME to
   the ID_TAG_PATH of the GPU that they want to use, and they can
   easily query this information using "udevadm info". This would
   be the same of setting the "DRMDevice" registry option for the
   Wine Wayland driver, but it is a little less work.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>