winewayland.drv: Implement cross-process vkCreateWin32SurfaceKHR.
When the target window for a Win32 VkSurfaceKHR is not owned by the
current process, we don't have access to the corresponding wayland
surface to use as the target for creating the host VkSurfaceKHR. In such
cases create a dummy wl_surface and use it to allow VkSurfaceKHR to
conclude successfully.

Note that no content targeting this dummy wl_surface will ever become
visible, but we will fix this in upcoming commits when we introduce a
cross-process swapchain implementation that forwards images to the right
process for presentation.

For now the cross-process code path is not used (notice the constant
allowed_to_create_remote_swapchain in the code), because we still don't
have all the pieces to make the cross-process swapchain work. So we
won't even allow the dummy wl_surface to be created. When we put all the
pieces together, we'll drop this constant and allow the code to run.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>
Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>