winewayland.drv: Move windows to the origin of their main output.
To maintain some degree of consistency between the Wayland surface and
Windows window positioning, place top-level windows on the output
dictated by the compositor. We position the window at the origin of that
output to maximize the window area that is accessible by mouse events.

We only want the latest invocation of output positioning to be handled, so
we attach a serial to each posted message and discard obsolete messages.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>