winewayland.drv: Introduce mechanism to schedule thread callbacks
Introduce an internal driver mechanism to schedule and dispatch thread
callbacks, to be run after the specified time has elapsed.

Note that the current mechanism doesn't wake up threads that are
currently blocked waiting, in order to service any triggered callbacks.
This deficiency will be fixed in the next commit.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>